Move function getAreaObject
